Initial commit
The goal of the importbounce tool is to act as a Go import redirector
for AWS Lambda. This way you can redirect your Go imports without having
to constantly run a server. Probably best for if your packages aren't
used that much, otherwise the Lambda costs might be high. In that case
you can just use the server that Russ Cox wrote:
https://github.com/rsc/go-import-redirector