Use custom CloudFront distribution in front of API
Unfortunately the ongoing pain of setting up AWS is not yet over. As it
turns out, API Gateway does *not* automatically redirect HTTP requests
to HTTPS, meaning that when you type a domain into your browser
expecting to be redirected, it just fails. The proper serverless way to
do this is apparently through CloudFront.

This change was especially fun to write, since every change to a
CloudFront distribution takes maybe 20 minutes or so to propagate, and
CloudFormation waits for it to complete before the stack is considered
updated.

I'm now using the magical environment variable from the Lambda proxy
library to set the request host, since forwarding the Host header
through CloudFront caused me to get a CloudFront-generated 403 page. Not
sure if it's CloudFront or API Gateway getting mad here.

Also, the stack as-written can only be deployed to us-east-1 since
that's where CloudFront requires the certificate to be. I'm thinking
about letting people provide the ARN of a pre-existing ACM certificate
as a parameter, and not creating the certificate in that case, assuming
CloudFormation can do that.