Remove upx from build process
The startup time is on the order of 10s of milliseconds, which is
actually pretty decent in terms of the typical execution time I see on
Lambda, especially since the original intent was to try to make this
start up faster by reducing download size. Should have tested first,
ridiculous miss on my part.