Initial attempt at supporting Lambda Function URLs
As of April 2022, AWS Lambda supports a "function URLs" feature that
provides a simple HTTPS gateway at no additional cost, following the
same request and response schemas as API Gateway 2.0.

In theory, this change will allow us to eliminate the custom API gateway
that normally sits between CloudFront and the Lambda function,
simplifying the stack and reducing costs. It'll probably take a lot of
trial and error to work out all the new kinks, though. It'll also be
interesting to see how easy it is to update an existing deployment in
place, and how CloudFormation handles all the associated changes.