Create an explicit CloudFront cache policy
First, this is now the recommended way to configure CloudFront caching
behavior, rather than using the legacy parameters like DefaultTTL.

Second, this new configuration is more tightly tuned to the headers and
query strings we're interested in, in particular the Host header. If
this works, we can finally look at getting rid of the GO_API_HOST hack.