Update description for ConfigFileNoSSL
Yes, it is strictly less secure, but it's worth noting that a MITM on
AWS internal networks is probably unlikely.

Maybe it would be worthwhile to embed the AWS roots like I do in other
projects and try turning TLS back on. Maybe S3 has added TLS 1.3
support, or some other part of the process has gotten better?