Limit HTTP requests to GET and HEAD
I've been rolling through all of my projects to make them more strict
about which HTTP methods they accept, after the net/http.ServeMux
changes in Go 1.22 made me realize that this is a detail I've
persistently missed in many (though not all) situations.

In my own importbounce deployment (for go.alexhamlin.co itself),
CloudFront already restricts which HTTP methods can be used (with an
HTTP 403 instead of a 405), so this is more of a defense-in-depth
against Hyrum's Law for a future standard-HTTP importbounce deployment,
or maybe deployment on something like Cloud Run.